chore: Move News to info (refs #226)
Explicit better than implicit
